+++ app.rb
@@ -5,6 +5,7 @@ App.call ARGV do
 
   system :dev
   system :happy
+  system :net
 
   # Modes help you organize your application's behavior and settings.
   # Learn more: http://guides.lizarb.org/modes.html