+++ app/net/databases/redis_db.rb
@@ -1,0 +1,5 @@
+class RedisDb < Liza::RedisDb
+
+  # Set up your database controllers per the DSL in http://guides.lizarb.org/controllers/database.html
+
+end