+++ lib/net_system/net/controllers/database/redis_db.rb
@@ -1,0 +1,11 @@
+class NetSystem
+  class RedisDb < Database
+    set_adapter :redis
+
+    def now
+      array = adapter.now
+      Time.at array[0]
+    end
+
+  end
+end