+++ lib/net_system/net/panels/adapter_panel.rb
@@ -1,0 +1,5 @@
+class NetSystem
+  class AdapterPanel < Liza::Panel
+
+  end
+end